rename __utilities.hpp to __detail/__meta.hpp
