+++ include/__memory.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include <__utility.hpp>
+#include <__detail/__meta.hpp>
 #include <concepts.hpp>
 
 #include <atomic>
