+++ test/detail/test_utility.cpp
@@ -15,7 +15,7 @@
  */
 
 #include <catch2/catch.hpp>
-#include <__utility.hpp>
+#include <__detail/__meta.hpp>
 
 #include <optional>
